backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
python-jose
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
httpx
resend